Corrections in CI docker tests setup

+++ docker/entrypoints/ci-test.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-cp .ci-cache/files/.eslintcache .eslintcache
+if [[ -f ".ci-cache/files/.eslintcache" ]]
+then
+    cp .ci-cache/files/.eslintcache .eslintcache
+fi
 make install-test
 pytest --gherkin-terminal-reporter -vv --showlocals --cov=funnel
 coverage lcov -o coverage/funnel.lcov
